requirements.txt
streamlit>=1.50.0
mysql-connector-python>=9.1.0
python-dotenv>=1.0.0
pandas>=2.0.0
